+++ env-setup.sh
@@ -14,3 +14,7 @@ sudo apt install ros-humble-grid-map-rviz-plugin
 # install cyclone DDS (lower RAM-usage)
 sudo apt update
 sudo apt install -y ros-humble-rmw-cyclonedds-cpp
+
+
+sudo apt update
+sudo apt install ros-humble-image-transport-plugins