+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+## [unreleased]
+
+### 🚀 Features
+
+- Add ArucoDetectionEngine
 ## [1.3.2] - 2026-04-27
 
 ### 🚀 Features
 
+- Enhance FollowerEngine with improved target acquisition and navigation logic
 - Improve hypothesis management and pose matching in MappingEngine
 
 ### ⚙️ Miscellaneous Tasks