+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Add ArucoDetectionEngine
 - Add aruco marker topic
 - Enhance FollowerEngine with improved target acquisition and navigation logic
+
+### ⚙️ Miscellaneous Tasks
+
+- Tune camera matrix and coefficients
 ## [1.3.2] - 2026-04-27
 
 ### 🚀 Features